FlaskApi/requirements.txt
Authlib==0.15.4
beautifulsoup4==4.11.1
dnspython==2.2.1
Flask==2.0.3
Flask_APScheduler==1.11.0
Flask_Cors==3.0.10
lxml==4.8.0
numpy==1.19.5
pandas==1.1.5
pyhanlp==0.1.84
PyMySQL==0.7.11
pypinyin==0.46.0
python_dateutil==2.8.2
requests==2.27.1
scikit_learn==0.23.2
xlwt==1.3.0
